feat: Setup changes, added postgres database connection, made models and migrated
This push switched setup so front and back are in original src for easier access, added connection with postgres database so as long as you have your postgres with the database you can use it locally, made all models and migrated it so data can be extracted for use. Need to add data to the database to check if it works.